Day 6: Working solution for Part Two

+++ Day/06/src/console/Domain/RaceRecord.cs
@@ -1,3 +1,3 @@
 namespace Domain;
 
-record RaceRecord(int Time, int Distance);
+record RaceRecord(long Time, long Distance);
